requirements.txt
python-dateutil==2.8.2
requests==2.26.0
Telethon==1.23.0
